test: refactored the test script to accepted standard
