--- test/unit/register/hybee-signup-test/.gitignore
@@ -1,1 +1,0 @@
-./node_modules
