--- test/unit/register/hybee-signup-test/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "signup-test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "signup.js",
-  "scripts": {
-    "test": "mocha"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@hapi/joi": "^17.1.1",
-    "axios": "^0.19.2",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1"
-  },
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "mocha": "^7.2.0"
-  }
-}
